adding support to observe events with ES7 decorators
splitting out the tests
