+++ examples/esp-chat-react-es6/js/model/ThreadSection.js
@@ -14,57 +14,48 @@ export default class ThreadSection extends esp.model.DisposableBase {
         this.hasChanges = false;
     }
     initialise() {
-        this._observeMessagesReceived();
-        this._observeThreadSelected();
+        this.addDisposable(this._router.observeEventsOn(this));
     }
     preProcess() {
         this.hasChanges = false;
     }
-    _observeMessagesReceived() {
-        var _this = this;
-        this.addDisposable(
-            this._router.getEventObservable('MessagesReceived').observe((model, event) => {
-                for (var i = 0; i < event.rawMessages.length; i++) {
-                    var rawMessage = event.rawMessages[i];
-                    var thread = _this.threadsById[rawMessage.threadId];
-                    var messageTime = new Date(rawMessage.timestamp);
-                    if (thread === undefined) {
-                        thread = new Thread(
-                            rawMessage.threadId,
-                            rawMessage.threadName,
-                            messageTime,
-                            rawMessage.text);
-                        _this.threadsById[rawMessage.threadId] = thread;
-                        _this.sortedThreads.push(thread);
-                    } else {
-                        if (thread.lastMessageTime <= messageTime) {
-                            thread.lastMessageTime = messageTime;
-                            thread.lastMessageText = rawMessage.text;
-                        }
-                    }
-                    if (thread.id === model.selectedThreadId) {
-                        thread.isRead = true;
-                    }
+    @esp.observeEvent('MessagesReceived')
+    _observeMessagesReceived(event, context, model) {
+        for (var i = 0; i < event.rawMessages.length; i++) {
+            var rawMessage = event.rawMessages[i];
+            var thread = _this.threadsById[rawMessage.threadId];
+            var messageTime = new Date(rawMessage.timestamp);
+            if (thread === undefined) {
+                thread = new Thread(
+                    rawMessage.threadId,
+                    rawMessage.threadName,
+                    messageTime,
+                    rawMessage.text);
+                _this.threadsById[rawMessage.threadId] = thread;
+                _this.sortedThreads.push(thread);
+            } else {
+                if (thread.lastMessageTime <= messageTime) {
+                    thread.lastMessageTime = messageTime;
+                    thread.lastMessageText = rawMessage.text;
                 }
-                _this.sortedThreads.sort(function (a, b) {
-                    return a.lastMessageTime > b.lastMessageTime ? -1 : a.lastMessageTime < b.lastMessageTime ? 1 : 0;
-                });
-                _this._updateActiveFlags(model);
-                _this._updateUnreadCount(model);
-                _this.hasChanges = true;
-            })
-        );
+            }
+            if (thread.id === model.selectedThreadId) {
+                thread.isRead = true;
+            }
+        }
+        _this.sortedThreads.sort(function (a, b) {
+            return a.lastMessageTime > b.lastMessageTime ? -1 : a.lastMessageTime < b.lastMessageTime ? 1 : 0;
+        });
+        _this._updateActiveFlags(model);
+        _this._updateUnreadCount(model);
+        _this.hasChanges = true;
     };
-    _observeThreadSelected() {
-        var _this = this;
-        this.addDisposable(
-            this._router.getEventObservable('ThreadSelected', esp.ObservationStage.committed).observe((model) => {
-                _this.threadsById[model.selectedThreadId].isRead = true;
-                _this._updateActiveFlags(model);
-                _this._updateUnreadCount(model);
-                _this.hasChanges = true;
-            })
-        );
+    @esp.observeEvent('ThreadSelected', esp.ObservationStage.committed)
+    _observeThreadSelected(event, context, model) {
+        _this.threadsById[model.selectedThreadId].isRead = true;
+        _this._updateActiveFlags(model);
+        _this._updateUnreadCount(model);
+        _this.hasChanges = true;
     };
     _updateActiveFlags(model) {
         for (var i = 0; i < this.sortedThreads.length; i++) {
