+++ examples/esp-chat-react-es6/webpack-config.js
@@ -14,7 +14,7 @@ module.exports = [ {
         module: {
             loaders: [
                 // {test: /\.jsx?$/, exclude: /node_modules/, loader: 'babel-loader?experimental&optional=runtime'}
-                {test: /\.js?$/, exclude: /node_modules/, loader: 'babel-loader'}
+                {test: /\.js?$/, exclude: /node_modules/, loader: 'babel-loader?stage=0'}
             ]
         },
         devtool: 'source-map'
