+++ karma.conf.js
@@ -32,7 +32,7 @@ module.exports = function (config) {
             devtool: "inline-source-map",
             module: {
                 loaders: [
-                    {test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader?sourceMap=true'}
+                    {test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader?sourceMap=true&stage=0'}
                 ]
             }
         },
