+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "esp-js",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "Evented State Processor (ESP) adds specific processing workflow around changes to a model's state",
   "keywords": [
     "state processor",
