+++ src/index.js
@@ -20,4 +20,4 @@ exports.ObservationStage = require('./router/ObservationStage');
 exports.Router = require('./router/Router');
 exports.SingleModelRouter = require('./router/SingleModelRouter');
 exports.model = require('./model');
-
+exports.observeEvent = require('./decorators/observeEvent');
